% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_two_numbers.R
\name{add_two_numbers}
\alias{add_two_numbers}
\title{Function to add two numbers}
\usage{
add_two_numbers(num1, num2)
}
\arguments{
\item{num1}{numeric value}

\item{num2}{numeric value}
}
\value{
numeric value which is the sum of two numbers
}
\description{
Function to add two numbers
}
\examples{
add_two_numbers(7, 11)
}
