% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply_two_numbers.R
\name{multiply_two_numbers}
\alias{multiply_two_numbers}
\title{Function to multiply two numbers}
\usage{
multiply_two_numbers(num1, num2)
}
\arguments{
\item{num1}{numeric value}

\item{num2}{numeric value}
}
\value{
numeric value which is the multiple of two numbers
}
\description{
Function to multiply two numbers
}
\examples{
multiply_two_numbers(7, 11)
}
