% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_two_numbers.R
\name{subtract_two_numbers}
\alias{subtract_two_numbers}
\title{Function to subtract two numbers}
\usage{
subtract_two_numbers(num1, num2)
}
\arguments{
\item{num1}{numeric value}

\item{num2}{numeric value}
}
\value{
numeric value which is the subtraction of the second number from the first number
}
\description{
Function to subtract two numbers
}
\examples{
subtract_two_numbers(7, 11)
}
